Move login logic to LoginUseCase
